requirements.txt
APScheduler==3.8.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==4.0.1
django-apscheduler==0.6.0
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
future==0.18.2
google==3.0.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.0
praw==7.5.0
prawcore==2.3.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tweepy==4.5.0
tzdata==2021.5
tzlocal==4.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.8
websocket-client==1.2.3
Werkzeug==2.0.3
